requirements.txt
streamlit==1.32.0
requests==2.31.0
pypdf2==3.0.1
transformers==4.40.0
torch==2.2.0
accelerate==0.29.0
sentencepiece==0.2.0
python-dotenv